gcp/function_code/requirements.txt
requests
functions_framework
google-cloud-bigquery